backend/fixieNotification/tests/requirements.txt
pytest
pytest-django
pytest-asyncio
hypothesis
freezegun
pytest-mock
channels
